Extract i2c into separate module
